rename: apiSlice 네이밍 통일
