+++ src/components/OnEul/index.css
@@ -14,4 +14,6 @@
 .item {
   border-top: 2px solid var(--primary);
   border-bottom: 2px solid var(--primary);
+  width: calc((100vw - 4px) / 2);
+  height: calc((100vw - 4px) / 2);
 }
