+++ src/store/api/productApiSlice.js
@@ -1,4 +1,4 @@
-import { apiSlice } from '../api/apiSlice'
+import { apiSlice } from './apiSlice'
 
 export const productApi = apiSlice.injectEndpoints({
   endpoints: (builder) => ({
