--- src/store/api/qeustions.js
@@ -1,75 +1,0 @@
-export const questions = [
-  {
-    id: 1,
-    createdDate: '2022.09.25',
-    answerYn: true,
-    type: '교환문의',
-    product: {
-      thumbnail:
-        'https://img.29cm.co.kr/next-product/2021/03/05/d22e8b4eb5fd4d52aa32a65b29b3c25f_20210305180022.jpg?width=300',
-      brand: 'TGT',
-      product: 'TGT PAISLEY HAIR BAND',
-      price: 1100,
-      count: 1,
-    },
-    images: [
-      'https://ifh.cc/g/YxCr1R.jpg',
-      'https://img.29cm.co.kr/next-product/2022/09/26/1d5795536cf247ae9cd2fd577eafbf6d_20220926160037.jpeg?width=800',
-    ],
-    title: '가방에 구멍이 나 있어서 다른 상품으로 교환하고 싶습니다',
-    content:
-      '가방에 구멍이 나 있어서 다른 상품으로 교환하고 싶습니다. 빠른 처리 부탁드려요. 가방에 구멍이 나 있어서 다른 상품으로 교환하고 싶습니다. 빠른 처리 부탁드려요. 가방에 구멍이 나 있어서 다른 상품으로 교환하고 싶습니다. 빠른 처리 부탁드려요.',
-    answer:
-      '안녕하세요. 000(아이디)님\n내 손 안의 매거진, 오늘의 쇼텐입니다.\n\n브랜드 측에서 확인하고 최대한 빠르게 교환 해드리도록 하겠습니다. \n\n감사합니다 :) ',
-    privateYn: '공개',
-    password: null,
-  },
-  {
-    id: 2,
-    createdDate: '2022.09.25',
-    answerYn: false,
-    type: '배송문의',
-    product: {
-      thumbnail: 'http://asq.kr/XHkPC05q4U',
-      brand: 'KUME',
-      product: 'TWEED BOMBER JACKET, BLACK',
-      price: 1100,
-      count: 1,
-    },
-    images: [
-      'https://image.wconcept.co.kr/productimg/image/img0/45/301852145_add1_GQ87761.jpg',
-    ],
-    title: '이번 주 안에 받아보고 싶은데 가능할까요?',
-    content: '이번 주 안에 받아보고 싶은데 가능할까요?',
-    answer: null,
-    privateYn: '비공개',
-    password: 1234,
-  },
-  {
-    id: 3,
-    createdDate: '2022.10.25',
-    answerYn: true,
-    type: '시스템문의',
-    title: '자꾸 로그인이 안되어요...',
-    product: null,
-    images: null,
-    content: '자꾸 로그인이 안되는데 확인 한번 부탁드립니다.',
-    answer:
-      '안녕하세요 고객님!\n\n오늘의 쇼핑 쇼텐입니다. \n\n말씀하신 부분에 있어서 정확한 확인을 위해 유선 안내를 드리고자 합니다. \n\n혹시 연락가능한 시간대가 있으실까요?',
-    privateYn: '공개',
-    password: null,
-  },
-  {
-    id: 4,
-    createdDate: '2022.10.25',
-    answerYn: false,
-    type: '이벤트문의',
-    title: '메인이벤트 관련해서 질문드려요~',
-    product: null,
-    images: null,
-    content: '이벤트문의관련 해서 질문드리니다. 어쩌구저쩌구 저어어어쩌구',
-    answer: null,
-    privateYn: '비공개',
-    password: 5678,
-  },
-]
