+++ src/store/api/questionApiSlice.js
@@ -1,4 +1,4 @@
-import { apiSlice } from '../api/apiSlice'
+import { apiSlice } from './apiSlice'
 
 export const questionApi = apiSlice.injectEndpoints({
   endpoints: (builder) => ({
